Code refactor with createCampaign API.
